+++ formguard/contrib/turnstile/widgets.py
@@ -6,17 +6,19 @@ class TurnstileWidget(forms.Widget):
 
     template_name = 'formguard/contrib/turnstile/widget.html'
 
-    def __init__(self, site_key, theme='auto', size='normal', attrs=None):
+    def __init__(self, site_key, theme='auto', size='normal', callback=None, attrs=None):
         super().__init__(attrs=attrs)
         self.site_key = site_key
         self.theme = theme
         self.size = size
+        self.callback = callback
 
     def get_context(self, name, value, attrs):
         context = super().get_context(name, value, attrs)
         context['widget']['site_key'] = self.site_key
         context['widget']['theme'] = self.theme
         context['widget']['size'] = self.size
+        context['widget']['callback'] = self.callback
         return context
 
     def value_from_datadict(self, data, files, name):